refactor(frontend): move pay method up to wizard
